test : add login screen & register screen unit test code
